get and default get not implemented more cleanly
